refactor(publish-flow): extract progress-sync into a mixin (Task 6)
Move sync_publish_progress / sync_scale_progress / sync_restart_progress plus
their SUCCESS/FAILURE handlers, device-binding update, and get_baas_publish_progress
into ProgressSyncMixin (verbatim; shares self with the facade — same behavior,
same test interception). Facade 2685 -> 2148 lines. 560 tests green.

Chose the mixin pattern (already used by publish_rollback_mixin.py) over a
context-object runner for the large verbatim method groups: lower risk, no body
rewrites.
